Remove core-js dependency.
Similarly to https://github.com/twada/call-matcher/pull/2,
core-js accounts for the vast majority of the install size of
espurify.

Most of these functions are supported in recent Node versions (from the
past few years), so it might be worth it to drop the ponyfills.

Map and Object.assign are only supported in Node 4 and up. The
Object.assign call was not strictly necessary (it was used to clone an
object, but the original would not be modified by espurify), so I
removed it. I also removed the Map import here, so Node 4 is required
with this patch; that may be a breaking change so it may be better to
depend on a (different) polyfill.

Lmk if the above is ok with you (and whether this is something you're
open to at all)